refactor(minecraft): extract version comparison, log filtering, and skin management from ModMinecraft
