+++ Plain Craft Launcher 2/Modules/Minecraft/ModMinecraft.cs
@@ -19,8 +19,6 @@ namespace PCL;
 
 public static class ModMinecraft
 {
-    public const string UNKNOWN_VERSION_KEY = "UnknownVersion";
-
     /// <summary>
     ///     发送 Minecraft 更新提示。
     /// </summary>
@@ -62,140 +60,6 @@ public static class ModMinecraft
         catch (Exception ex)
         {
             ModBase.Log(ex, Lang.Text("Minecraft.Error.UpdateNotify", versionName ?? "Nothing"), ModBase.LogLevel.Feedback);
-        }
-    }
-
-    /// <summary>
-    ///     比较两个版本名；等同 Left >= Right。
-    ///     无法比较两个预发布版的大小。
-    ///     支持的格式：未知版本, 1.13.2, 1.7.10-pre4, 1.8_pre, 1.14 Pre-Release 2, 1.14.4 C6
-    /// </summary>
-    public static bool CompareVersionGe(string left, string right)
-    {
-        return CompareVersion(left, right) >= 0;
-    }
-
-    /// <summary>
-    ///     比较两个版本名，若 Left 较新则返回 1，相同则返回 0，Right 较新则返回 -1；等同 Left - Right。
-    ///     无法比较两个预发布版的大小。
-    ///     支持的格式：未知版本, 26.1-snapshot-1，1.13.2, 1.7.10-pre4, 1.8_pre, 1.14 Pre-Release 2, 1.14.4 C6
-    /// </summary>
-    public static int CompareVersion(string left, string right)
-    {
-        if (left == Lang.Text("Minecraft.Version.Unknown") || right == Lang.Text("Minecraft.Version.Unknown"))
-        {
-            if (left == Lang.Text("Minecraft.Version.Unknown") && right != Lang.Text("Minecraft.Version.Unknown"))
-                return 1;
-            if (left == Lang.Text("Minecraft.Version.Unknown") && right == Lang.Text("Minecraft.Version.Unknown"))
-                return 0;
-            if (left != Lang.Text("Minecraft.Version.Unknown") && right == Lang.Text("Minecraft.Version.Unknown"))
-                return -1;
-        }
-
-        left = left.ToLowerInvariant();
-        right = right.ToLowerInvariant();
-        var lefts = left.Replace("快照", "snapshot").Replace("预览版", "pre").RegexSearch("[a-z]+|[0-9]+");
-        var rights = right.Replace("快照", "snapshot").Replace("预览版", "pre").RegexSearch("[a-z]+|[0-9]+");
-        var i = 0;
-        while (true)
-        {
-            // 两边均缺失，感觉是一个东西
-            if (lefts.Count - 1 < i && rights.Count - 1 < i)
-            {
-                if (string.CompareOrdinal(left, right) > 0)
-                    return 1;
-                if (string.CompareOrdinal(left, right) < 0)
-                    return -1;
-                return 0;
-            }
-
-            // 确定两边的数值
-            var leftValue = lefts.Count - 1 < i ? "0" : lefts[i];
-            var rightValue = rights.Count - 1 < i ? "0" : rights[i];
-            if ((leftValue ?? "") == (rightValue ?? ""))
-                goto NextEntry;
-            if (leftValue == "rc")
-                leftValue = (-1).ToString();
-            if (leftValue == "pre")
-                leftValue = (-2).ToString();
-            if (leftValue == "snapshot")
-                leftValue = (-3).ToString();
-            if (leftValue == "experimental")
-                leftValue = (-4).ToString();
-            var leftValValue = ModBase.Val(leftValue);
-            if (rightValue == "rc")
-                rightValue = (-1).ToString();
-            if (rightValue == "pre")
-                rightValue = (-2).ToString();
-            if (rightValue == "snapshot")
-                rightValue = (-3).ToString();
-            if (rightValue == "experimental")
-                rightValue = (-4).ToString();
-            var rightValValue = ModBase.Val(rightValue);
-            if (leftValValue == 0d && rightValValue == 0d)
-            {
-                // 如果没有数值则直接比较字符串
-                if (string.CompareOrdinal(leftValue, rightValue) > 0) return 1;
-
-                if (string.CompareOrdinal(leftValue, rightValue) < 0) return -1;
-            }
-            // 如果有数值则比较数值
-            // 这会使得一边是数字一边是字母时数字方更大
-            else if (leftValValue > rightValValue)
-            {
-                return 1;
-            }
-            else if (leftValValue < rightValValue)
-            {
-                return -1;
-            }
-
-            NextEntry: ;
-
-            i += 1;
-        }
-
-        return 0;
-    }
-
-    /// <summary>
-    ///     打码字符串中的 AccessToken。
-    /// </summary>
-    public static string FilterAccessToken(string raw, char filterChar)
-    {
-        // 打码 "accessToken " 后的内容
-        if (raw.Contains("accessToken "))
-            foreach (var Token in raw.RegexSearch("(?<=accessToken ([^ ]{5}))[^ ]+(?=[^ ]{5})"))
-                raw = raw.Replace(Token, new string(filterChar, Token.Count()));
-        // 打码当前登录的结果
-        var accessToken = ModLaunch.mcLoginLoader.output.AccessToken;
-        if (accessToken is not null && accessToken.Length >= 10 && raw.ContainsF(accessToken, true) &&
-            (ModLaunch.mcLoginLoader.output.Uuid ?? "") !=
-            (ModLaunch.mcLoginLoader.output.AccessToken ?? "")) // UUID 和 AccessToken 一样则不打码
-            raw = raw.Replace(accessToken, accessToken[..5] + new string(filterChar, accessToken.Length - 10) +
-                                           accessToken[^5..]);
-        return raw;
-    }
-
-    /// <summary>
-    ///     打码字符串中的 Windows 用户名。
-    /// </summary>
-    public static string FilterUserName(string raw, char filterChar)
-    {
-        var userProfile = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-        var userName = userProfile.Split(@"\").Last();
-        var maskedProfile = userProfile.Replace(userName, new string(filterChar, userName.Length));
-        return raw.Replace(userProfile, maskedProfile);
-    }
-
-    /// <summary>
-    ///     比较两个版本名的排序器。
-    /// </summary>
-    public class VersionComparer : IComparer<string>
-    {
-        public int Compare(string x, string y)
-        {
-            return CompareVersion(x, y);
         }
     }
 
@@ -1026,182 +890,7 @@ public static class ModMinecraft
 
     #endregion
 
-    #region 皮肤
 
-    public struct McSkinInfo
-    {
-        public bool IsSlim;
-        public string LocalFile;
-        public bool IsVaild;
-    }
-
-    /// <summary>
-    ///     要求玩家选择一个皮肤文件，并进行相关校验。
-    /// </summary>
-    public static McSkinInfo McSkinSelect()
-    {
-        var fileName = SystemDialogs.SelectFile(Lang.Text("Launch.Skin.FileDialog.Filter"), Lang.Text("Launch.Skin.FileDialog.Title"));
-
-        // 验证有效性
-        if (string.IsNullOrEmpty(fileName))
-            return new McSkinInfo { IsVaild = false };
-        try
-        {
-            var image = new MyBitmap(fileName);
-            if (image.pic.Width != 64 || !(image.pic.Height == 32 || image.pic.Height == 64))
-            {
-                ModMain.Hint(Lang.Text("Launch.Skin.InvalidSize"), ModMain.HintType.Critical);
-                return new McSkinInfo { IsVaild = false };
-            }
-
-            var fileInfo = new FileInfo(fileName);
-            if (fileInfo.Length > 24 * 1024)
-            {
-                ModMain.Hint(Lang.Text("Launch.Skin.FileTooLarge", Lang.Number(fileInfo.Length / 1024d, "N2")),
-                    ModMain.HintType.Critical);
-                return new McSkinInfo { IsVaild = false };
-            }
-        }
-        catch (Exception ex)
-        {
-            ModBase.Log(ex, Lang.Text("Launch.Skin.File.Error"), ModBase.LogLevel.Hint);
-            return new McSkinInfo { IsVaild = false };
-        }
-
-        // 获取皮肤种类
-        var isSlim = ModMain.MyMsgBox(Lang.Text("Launch.Skin.Model.SelectMessage"), Lang.Text("Launch.Skin.Model.SelectTitle"), Lang.Text("Launch.Skin.Model.Steve"), Lang.Text("Launch.Skin.Model.Alex"), Lang.Text("Common.Option.IDontKnow"),
-            highLight: false);
-        if (isSlim == 3)
-        {
-            ModMain.Hint(Lang.Text("Launch.Skin.Model.UnknownHint"));
-            return new McSkinInfo { IsVaild = false };
-        }
-
-        return new McSkinInfo { IsVaild = true, IsSlim = isSlim == 2, LocalFile = fileName };
-    }
-
-    /// <summary>
-    ///     获取 Uuid 对应的皮肤文件地址，失败将抛出异常。
-    /// </summary>
-    public static string McSkinGetAddress(string uuid, string type)
-    {
-        if (string.IsNullOrEmpty(uuid))
-            throw new Exception(Lang.Text("Minecraft.Skin.Error.UuidEmpty"));
-
-        if (uuid.StartsWith("00000"))
-            throw new Exception(Lang.Text("Minecraft.Skin.Error.OfflineNoSkin"));
-
-        // 尝试读取缓存
-        var cachePath = Path.Combine(ModBase.pathTemp, $"Cache\\Skin\\Index{type}.ini");
-        var cacheSkinAddress = ModBase.ReadIni(cachePath, uuid);
-        if (!string.IsNullOrEmpty(cacheSkinAddress))
-            return cacheSkinAddress;
-
-        // 获取皮肤地址
-        var url = type switch
-        {
-            "Mojang" => "https://sessionserver.mojang.com/session/minecraft/profile/",
-            "Ms" => "https://sessionserver.mojang.com/session/minecraft/profile/",
-            "Auth" => ModProfile.selectedProfile.Server.Replace("/authserver", "") +
-                      "/sessionserver/session/minecraft/profile/",
-            _ => throw new ArgumentException(Lang.Text("Minecraft.Skin.Error.InvalidSkinType", type ?? "null"))
-        };
-
-        var skinString = ModNet.NetGetCodeByRequestRetry(url + uuid);
-        if (string.IsNullOrEmpty((string?)skinString))
-            throw new Exception(Lang.Text("Minecraft.Skin.Error.SkinReturnEmpty"));
-
-        // 解析皮肤 Property
-        string skinValue = null;
-        try
-        {
-            var json = (JsonObject)ModBase.GetJson((string)skinString);
-            foreach (var property in json["properties"].AsArray())
-                if (property["name"]?.ToString() == "textures")
-                {
-                    skinValue = property["value"]?.ToString();
-                    break;
-                }
-
-            if (skinValue is null)
-                throw new Exception(Lang.Text("Minecraft.Skin.Error.PropertyNotFound"));
-        }
-        catch (Exception ex)
-        {
-            ModBase.Log(ex,
-                $"无法完成解析的皮肤返回值，可能是未设置自定义皮肤的用户：{skinString}",
-                ModBase.LogLevel.Developer);
-            throw new Exception(Lang.Text("Minecraft.Skin.Error.NoSkinData"), ex);
-        }
-
-        // 解码 Base64 并解析 JSON
-        var decoded = Encoding.UTF8.GetString(Convert.FromBase64String(skinValue));
-        var skinJson = (JsonObject)ModBase.GetJson(decoded.ToLowerInvariant());
-
-        if (skinJson["textures"]?["skin"]?["url"] is null)
-            throw new Exception(Lang.Text("Minecraft.Skin.Error.NoCustomSkin"));
-
-        var skinUrl = skinJson["textures"]["skin"]["url"].ToString();
-        skinUrl = skinUrl.Contains("minecraft.net/") ? skinUrl.Replace("http://", "https://") : skinUrl;
-
-        // 保存缓存
-        ModBase.WriteIni(cachePath, uuid, skinUrl);
-        ModBase.Log($"[Skin] UUID {uuid} 对应的皮肤文件为 {skinUrl}");
-
-        return skinUrl;
-    }
-
-    private static readonly object mcSkinDownloadLock = new();
-
-    /// <summary>
-    ///     从 Url 下载皮肤。返回本地文件路径，失败将抛出异常。
-    /// </summary>
-    public static string McSkinDownload(string address)
-    {
-        var skinName = ModBase.GetFileNameFromPath(address);
-        var fileAddress = ModBase.pathTemp + @"Cache\Skin\" + ModBase.GetHash(address) + ".png";
-        lock (mcSkinDownloadLock)
-        {
-            if (!File.Exists(fileAddress))
-            {
-                FileDownloader.Download(address, fileAddress + ModNet.netDownloadEnd).GetAwaiter().GetResult();
-                File.Delete(fileAddress);
-                FileSystem.Rename(fileAddress + ModNet.netDownloadEnd, fileAddress);
-                ModBase.Log("[Minecraft] 皮肤下载成功：" + fileAddress);
-            }
-
-            return fileAddress;
-        }
-    }
-
-    /// <summary>
-    ///     获取 Uuid 对应的皮肤，返回“Steve”或“Alex”。
-    /// </summary>
-    public static string McSkinSex(string uuid)
-    {
-        if (uuid.Length != 32)
-            return "Steve";
-        var a = int.Parse(uuid[7].ToString(), NumberStyles.AllowHexSpecifier);
-        var b = int.Parse(uuid[15].ToString(), NumberStyles.AllowHexSpecifier);
-        var c = int.Parse(uuid[23].ToString(), NumberStyles.AllowHexSpecifier);
-        var d = int.Parse(uuid[31].ToString(), NumberStyles.AllowHexSpecifier);
-        return ((a ^ b ^ c ^ d) % 2) != 0 ? "Alex" : "Steve";
-        // Math.floorMod(uuid.hashCode(), 18)
-
-        // Public Function hashCode(ByVal str As String) As Integer
-        // Dim hash As Integer = 0
-        // Dim n As Integer = str.Length
-        // If n = 0 Then
-        // Return hash
-        // End If
-        // For i As Integer = 0 To n - 1
-        // hash = hash + Asc(str(i)) * (1 << (n - i - 1))
-        // Next
-        // Return hash
-        // End Function
-    }
-
-    #endregion
 
     #region 支持库文件（Libraries）
 
@@ -1563,7 +1252,7 @@ public static class ModMinecraft
                 {
                     ModBase.Log(
                         $"[Minecraft] 发现重复的支持库：{basicArray[i]} ({basicArrayVersion}) 与 {resultArray[key]} ({resultArrayVersion})，已忽略其中之一");
-                    if (CompareVersionGe(basicArrayVersion, resultArrayVersion)) resultArray[key] = basicArray[i];
+                    if (McVersionComparer.CompareVersionGe(basicArrayVersion, resultArrayVersion)) resultArray[key] = basicArray[i];
                 }
             }
             else
